requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-multipart
jinja2
